pyproject.toml
[tool.poetry]
name = "Diploma project on test automation in Python + Selene"
version = "0.1.0"
description = "UI + API + Mobile(browserstack) testing with selene, appium, pytest, request and allure( allure testops, allure)+ with notification(slack+telegram)"
authors = ["Bohdan Obruch <bodan1992@gmail.com>"]


[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.3.1"
selene = {version = "^2.0.0-beta.14", allow-prereleases = true}
allure-pytest = "^2.13.2"
Appium-Python-Client = "^2.10.0"
pytest-xdist = "^2.5.0"
requests = "^2.30.0"
pydantic = "^1.10.2"
lorem = "0.1.1"
python-dotenv = "^1.0.0"
voluptuous = "^0.13.1"
pytest-voluptuous = "^1.2.0"
curlify = "^2.2.1"
selenium = ">=4.9.1"
Faker = "^15.1.3"
allure-python-commons = "^2.12.0"
webdriver-manager = "^3.8.5"

[tool.poetry.dev-dependencies]
pylint = "^2.16.0"

[tool.pytest.ini_options]
addopts="-vv --alluredir=allure-results --clean-alluredir"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
